+++ src/Boundary/SphericalBoundary.cc
@@ -169,7 +169,7 @@ setGhostNodes(NodeList<Dim<3> >& nodeList) {
 
   // Create the set of ghost nodes on the NodeList.
   const int numNewGhostNodes = ghostNodes.size();
-  CHECK(controlNodes.size() == numNewGhostNodes);
+  CHECK((int)controlNodes.size() == numNewGhostNodes);
   nodeList.numGhostNodes(currentNumGhostNodes + numNewGhostNodes);
 
   // Use the updateGhostNodes method to set the node positions and H's.
@@ -209,8 +209,8 @@ updateGhostNodes(NodeList<Dim<3> >& nodeList) {
          ++ghostPositionItr, ++controlItr, ++ghostItr) {
       CHECK(controlItr < controlNodes.end());
       CHECK(ghostItr < ghostNodes.end());
-      CHECK(*controlItr == i);
-      CHECK(*ghostItr >= nodeList.firstGhostNode() && *ghostItr < nodeList.numNodes());
+      CHECK(*controlItr == (int)i);
+      CHECK(*ghostItr >= (int)nodeList.firstGhostNode() && *ghostItr < (int)nodeList.numNodes());
       const double rj = ghostPositionItr->magnitude();
       *ghostPositionItr *= ri*rj/(rj*rj + 1.0e-50);
       position(*ghostItr) = *ghostPositionItr;
@@ -243,8 +243,8 @@ applyGhostBoundary(Field<Dim<3> , int>& field) const {
   vector<int>::const_iterator ghostItr = ghostBegin(nodeList);
   for (; controlItr < controlEnd(nodeList); ++controlItr, ++ghostItr) {
     CHECK(ghostItr < ghostEnd(nodeList));
-    CHECK(*controlItr >= 0 && *controlItr < nodeList.numNodes());
-    CHECK(*ghostItr >= nodeList.firstGhostNode() && *ghostItr < nodeList.numNodes());
+    CHECK(*controlItr >= 0 && *controlItr < (int)nodeList.numNodes());
+    CHECK(*ghostItr >= (int)nodeList.firstGhostNode() && *ghostItr < (int)nodeList.numNodes());
     field(*ghostItr) = field(*controlItr);
   }
 }
@@ -261,8 +261,8 @@ applyGhostBoundary(Field<Dim<3> , Dim<3>::Scalar>& field) const {
   vector<int>::const_iterator ghostItr = ghostBegin(nodeList);
   for (; controlItr < controlEnd(nodeList); ++controlItr, ++ghostItr) {
     CHECK(ghostItr < ghostEnd(nodeList));
-    CHECK(*controlItr >= 0 && *controlItr < nodeList.numNodes());
-    CHECK(*ghostItr >= nodeList.firstGhostNode() && *ghostItr < nodeList.numNodes());
+    CHECK(*controlItr >= 0 && *controlItr < (int)nodeList.numNodes());
+    CHECK(*ghostItr >= (int)nodeList.firstGhostNode() && *ghostItr < (int)nodeList.numNodes());
     field(*ghostItr) = field(*controlItr);
   }
 }
@@ -280,8 +280,8 @@ applyGhostBoundary(Field<Dim<3>, Dim<3>::Vector>& field) const {
   vector<int>::const_iterator ghostItr = ghostBegin(nodeList);
   for (; controlItr < controlEnd(nodeList); ++controlItr, ++ghostItr) {
     CHECK(ghostItr < ghostEnd(nodeList));
-    CHECK(*controlItr >= 0 && *controlItr < nodeList.numNodes());
-    CHECK(*ghostItr >= nodeList.firstGhostNode() && *ghostItr < nodeList.numNodes());
+    CHECK(*controlItr >= 0 && *controlItr < (int)nodeList.numNodes());
+    CHECK(*ghostItr >= (int)nodeList.firstGhostNode() && *ghostItr < (int)nodeList.numNodes());
     field(*ghostItr) = reflectOperator(position(*controlItr),
                                        position(*ghostItr))*field(*controlItr);
   }
@@ -300,8 +300,8 @@ applyGhostBoundary(Field<Dim<3>, Dim<3>::Tensor>& field) const {
   vector<int>::const_iterator ghostItr = ghostBegin(nodeList);
   for (; controlItr < controlEnd(nodeList); ++controlItr, ++ghostItr) {
     CHECK(ghostItr < ghostEnd(nodeList));
-    CHECK(*controlItr >= 0 && *controlItr < nodeList.numNodes());
-    CHECK(*ghostItr >= nodeList.firstGhostNode() && *ghostItr < nodeList.numNodes());
+    CHECK(*controlItr >= 0 && *controlItr < (int)nodeList.numNodes());
+    CHECK(*ghostItr >= (int)nodeList.firstGhostNode() && *ghostItr < (int)nodeList.numNodes());
     const Tensor T = reflectOperator(position(*controlItr), position(*ghostItr));
     field(*ghostItr) = T*(field(*controlItr)*T);
   }
@@ -320,8 +320,8 @@ applyGhostBoundary(Field<Dim<3>, Dim<3>::SymTensor>& field) const {
   vector<int>::const_iterator ghostItr = ghostBegin(nodeList);
   for (; controlItr < controlEnd(nodeList); ++controlItr, ++ghostItr) {
     CHECK(ghostItr < ghostEnd(nodeList));
-    CHECK(*controlItr >= 0 && *controlItr < nodeList.numNodes());
-    CHECK(*ghostItr >= nodeList.firstGhostNode() && *ghostItr < nodeList.numNodes());
+    CHECK(*controlItr >= 0 && *controlItr < (int)nodeList.numNodes());
+    CHECK(*ghostItr >= (int)nodeList.firstGhostNode() && *ghostItr < (int)nodeList.numNodes());
     const Tensor T = reflectOperator(position(*controlItr), position(*ghostItr));
     field(*ghostItr) = T*(field(*controlItr)*T).Symmetric();
   }
@@ -341,8 +341,8 @@ applyGhostBoundary(Field<Dim<3>, Dim<3>::ThirdRankTensor>& field) const {
   ThirdRankTensor val;
   for (; controlItr < controlEnd(nodeList); ++controlItr, ++ghostItr) {
     CHECK(ghostItr < ghostEnd(nodeList));
-    CHECK(*controlItr >= 0 && *controlItr < nodeList.numNodes());
-    CHECK(*ghostItr >= nodeList.firstGhostNode() && *ghostItr < nodeList.numNodes());
+    CHECK(*controlItr >= 0 && *controlItr < (int)nodeList.numNodes());
+    CHECK(*ghostItr >= (int)nodeList.firstGhostNode() && *ghostItr < (int)nodeList.numNodes());
     const Tensor T = reflectOperator(position(*controlItr), position(*ghostItr));
     val = ThirdRankTensor::zero;
     const ThirdRankTensor& fc = field(*controlItr);
@@ -377,8 +377,8 @@ applyGhostBoundary(Field<Dim<3>, Dim<3>::FourthRankTensor>& field) const {
   FourthRankTensor val;
   for (; controlItr < controlEnd(nodeList); ++controlItr, ++ghostItr) {
     CHECK(ghostItr < ghostEnd(nodeList));
-    CHECK(*controlItr >= 0 && *controlItr < nodeList.numNodes());
-    CHECK(*ghostItr >= nodeList.firstGhostNode() && *ghostItr < nodeList.numNodes());
+    CHECK(*controlItr >= 0 && *controlItr < (int)nodeList.numNodes());
+    CHECK(*ghostItr >= (int)nodeList.firstGhostNode() && *ghostItr < (int)nodeList.numNodes());
     const Tensor T = reflectOperator(position(*controlItr), position(*ghostItr));
     val = FourthRankTensor::zero;
     const FourthRankTensor& fc = field(*controlItr);
@@ -417,8 +417,8 @@ applyGhostBoundary(Field<Dim<3>, Dim<3>::FifthRankTensor>& field) const {
   FifthRankTensor val;
   for (; controlItr < controlEnd(nodeList); ++controlItr, ++ghostItr) {
     CHECK(ghostItr < ghostEnd(nodeList));
-    CHECK(*controlItr >= 0 && *controlItr < nodeList.numNodes());
-    CHECK(*ghostItr >= nodeList.firstGhostNode() && *ghostItr < nodeList.numNodes());
+    CHECK(*controlItr >= 0 && *controlItr < (int)nodeList.numNodes());
+    CHECK(*ghostItr >= (int)nodeList.firstGhostNode() && *ghostItr < (int)nodeList.numNodes());
     const Tensor T = reflectOperator(position(*controlItr), position(*ghostItr));
     val = FifthRankTensor::zero;
     const FifthRankTensor& fc = field(*controlItr);
@@ -460,8 +460,8 @@ applyGhostBoundary(Field<Dim<3>, Dim<3>::FacetedVolume>& field) const {
   auto ghostItr = ghostBegin(nodeList);
   for (; controlItr < controlEnd(nodeList); ++controlItr, ++ghostItr) {
     CHECK(ghostItr < ghostEnd(nodeList));
-    CHECK(*controlItr >= 0 && *controlItr < nodeList.numNodes());
-    CHECK(*ghostItr >= nodeList.firstGhostNode() && *ghostItr < nodeList.numNodes());
+    CHECK(*controlItr >= 0 && *controlItr < (int)nodeList.numNodes());
+    CHECK(*ghostItr >= (int)nodeList.firstGhostNode() && *ghostItr < (int)nodeList.numNodes());
     const auto T = reflectOperator(position(*controlItr), position(*ghostItr));
     const auto& poly = field(*controlItr);
     vector<Vector> verts(poly.vertices());
@@ -483,8 +483,8 @@ applyGhostBoundary(Field<Dim<3> , std::vector<Dim<3>::Scalar> >& field) const {
   vector<int>::const_iterator ghostItr = ghostBegin(nodeList);
   for (; controlItr < controlEnd(nodeList); ++controlItr, ++ghostItr) {
     CHECK(ghostItr < ghostEnd(nodeList));
-    CHECK(*controlItr >= 0 && *controlItr < nodeList.numNodes());
-    CHECK(*ghostItr >= nodeList.firstGhostNode() && *ghostItr < nodeList.numNodes());
+    CHECK(*controlItr >= 0 && *controlItr < (int)nodeList.numNodes());
+    CHECK(*ghostItr >= (int)nodeList.firstGhostNode() && *ghostItr < (int)nodeList.numNodes());
     field(*ghostItr) = field(*controlItr);
   }
 }
@@ -564,7 +564,7 @@ enforceBoundary(Field<Dim<3>, Dim<3>::Vector>& field) const {
   for (vector<int>::const_iterator itr = violationBegin(nodeList);
        itr != violationEnd(nodeList);
        ++itr) {
-    CHECK(*itr >= 0 && *itr < nodeList.numInternalNodes());
+    CHECK(*itr >= 0 && *itr < (int)nodeList.numInternalNodes());
     field(*itr).y(0.0);
     field(*itr).z(0.0);
   }
@@ -580,7 +580,7 @@ enforceBoundary(Field<Dim<3>, Dim<3>::Tensor>& field) const {
   for (vector<int>::const_iterator itr = violationBegin(nodeList);
        itr != violationEnd(nodeList);
        ++itr) {
-    CHECK(*itr >= 0 && *itr < nodeList.numInternalNodes());
+    CHECK(*itr >= 0 && *itr < (int)nodeList.numInternalNodes());
     const double xx = field(*itr).xx();
     const double yy = field(*itr).yy();
     const double zz = field(*itr).zz();
@@ -600,7 +600,7 @@ enforceBoundary(Field<Dim<3>, Dim<3>::SymTensor>& field) const {
   for (vector<int>::const_iterator itr = violationBegin(nodeList);
        itr != violationEnd(nodeList);
        ++itr) {
-    CHECK(*itr >= 0 && *itr < nodeList.numInternalNodes());
+    CHECK(*itr >= 0 && *itr < (int)nodeList.numInternalNodes());
     const double xx = field(*itr).xx();
     const double yy = field(*itr).yy();
     const double zz = field(*itr).zz();
