+++ src/Utilities/peanoHilbertOrderIndices.cc
@@ -95,8 +95,8 @@ hashPosition(const Dim<2>::Vector& position,
   const Vector cellsize = boxsize/ncells;
   const int xfine = int(deltar.x()/cellsize.x());
   const int yfine = int(deltar.y()/cellsize.y());
-  CHECK(xfine >= 0 and xfine < ncells);
-  CHECK(yfine >= 0 and yfine < ncells);
+  CHECK(xfine >= 0 and xfine < (int)ncells);
+  CHECK(yfine >= 0 and yfine < (int)ncells);
 
   // Recursively quadrant the position, until we get to the desired level.
   for (auto level = 0u; level != KeyTraits::numbits1d; ++level) {
@@ -108,8 +108,8 @@ hashPosition(const Dim<2>::Vector& position,
 
     const Key ncells = KeyTraits::two << level;
     CONTRACT_VAR(ncells);
-    CHECK(x >= 0 and x < ncells);
-    CHECK(y >= 0 and y < ncells);
+    CHECK(x >= 0 and x < (int)ncells);
+    CHECK(y >= 0 and y < (int)ncells);
 
     // Decide which (lab frame) quadrant this position represents.
     const int xx = x % 2;
@@ -203,9 +203,9 @@ hashPosition(const Dim<3>::Vector& position,
   const int xfine = int(deltar.x()/cellsize.x());
   const int yfine = int(deltar.y()/cellsize.y());
   const int zfine = int(deltar.z()/cellsize.z());
-  CHECK(xfine >= 0 and xfine < ncells);
-  CHECK(yfine >= 0 and yfine < ncells);
-  CHECK(zfine >= 0 and zfine < ncells);
+  CHECK(xfine >= 0 and xfine < (int)ncells);
+  CHECK(yfine >= 0 and yfine < (int)ncells);
+  CHECK(zfine >= 0 and zfine < (int)ncells);
 
   // Recursively quadrant the position, until we get to the desired level.
   for (auto level = 0u; level != KeyTraits::numbits1d; ++level) {
@@ -218,9 +218,9 @@ hashPosition(const Dim<3>::Vector& position,
 
     const Key ncells = KeyTraits::two << level;
     CONTRACT_VAR(ncells);
-    CHECK(x >= 0 and x < ncells);
-    CHECK(y >= 0 and y < ncells);
-    CHECK(z >= 0 and z < ncells);
+    CHECK(x >= 0 and x < (int)ncells);
+    CHECK(y >= 0 and y < (int)ncells);
+    CHECK(z >= 0 and z < (int)ncells);
 
     // Decide which (lab frame) quadrant this position represents.
     const int xx = x % 2;
